~ fixing ghost players ~ fixing start with shield

+++ ui/ClientUI.hpp
@@ -18,12 +18,12 @@ struct UISelect {
     char nameBuffer[31] = "";
     int selectedPreset = -1;
     int selectedColor = 0;
-    int selectedWeapon = 1;
+    short selectedWeapon = 1;
 
     bool allow_moove;
 
     int selectedZone = 0; // 0=NAME, 1=COLOR, 2=WEAPON, 3=CONFIRM
-    int subSelect = 0;    // 0=left, 1=right (ou bouton unique pour confirm)
+    int subSelect = 0;    // 0=left, 1=right (or unic for confirm)
 };
 
 class ClientUI : public Client {
